adding support for echo legacy to credit calculation
