+++ lib/holiday-stops/src/test/scala/com/gu/holiday_stops/subscription/SubscriptionDataIntegrationTest.scala
@@ -2,56 +2,16 @@ package com.gu.holiday_stops.subscription
 
 import java.time.LocalDate
 import java.time.temporal.ChronoUnit
+
 import com.gu.holiday_stops.Fixtures
-import org.scalatest.{EitherValues, FlatSpec, Matchers}
+import org.scalatest.EitherValues._
+import org.scalatest.Matchers._
 
-class SubscriptionDataIntegrationTest extends FlatSpec with Matchers with EitherValues {
-  MutableCalendar.setFakeToday(Some(LocalDate.of(2019, 10, 4)))
-
-  "SubscriptionDataIntegrationTest" should "calculate issue data correctly for GW 6 for 6" in {
-    val subscriptionFile = "GuardianWeeklyWith6For6.json"
-    val startDate = LocalDate.parse("2019-10-04")
-    val sixForSixBillingPeriod = BillingPeriod( startDate, startDate.plusWeeks(6).minusDays(1))
-    val normalBillingPeriod = BillingPeriod(
-      startDate.plusWeeks(6),
-      startDate.plusWeeks(6).plusMonths(3).minusDays(1)
-    )
-    val normalBillingPeriod2 = BillingPeriod(
-      startDate.plusWeeks(6).plusMonths(3),
-      startDate.plusWeeks(6).plusMonths(6).minusDays(1)
-    )
-
-    val expectedIssueData = List(
-      IssueData(startDate, sixForSixBillingPeriod, -1),
-      IssueData(startDate.plusWeeks(1), sixForSixBillingPeriod, -1),
-      IssueData(startDate.plusWeeks(2), sixForSixBillingPeriod, -1),
-      IssueData(startDate.plusWeeks(3), sixForSixBillingPeriod, -1),
-      IssueData(startDate.plusWeeks(4), sixForSixBillingPeriod, -1),
-      IssueData(startDate.plusWeeks(5), sixForSixBillingPeriod, -1),
-      IssueData(startDate.plusWeeks(6), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(7), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(8), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(9), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(10), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(11), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(12), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(13), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(14), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(15), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(16), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(17), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(18), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(19), normalBillingPeriod, -2.89),
-      IssueData(startDate.plusWeeks(20), normalBillingPeriod2, -2.89),
-    )
-
-    testSubscriptonDataIssueGeneration(subscriptionFile, startDate, expectedIssueData)
-  }
-
-  private def testSubscriptonDataIssueGeneration(subscriptionFile: String, startDate: LocalDate, expectedIssueData: List[IssueData]) = {
+object SubscriptionDataIntegrationTest {
+  def testSubscriptonDataIssueGeneration(subscriptionFile: String, startDate: LocalDate, expectedIssueData: List[IssueData]) = {
     val subscription = Fixtures.subscriptionFromJson(subscriptionFile)
     val subscriptionData = SubscriptionData(subscription).right.value
-    val datesToTest = getDatesToTest(expectedIssueData)
+    val datesToTest = getDatesToTest(startDate, expectedIssueData)
 
     subscriptionData.issueDataForDate(startDate.minusDays(1)).isLeft should equal(true)
 
@@ -66,9 +26,9 @@ class SubscriptionDataIntegrationTest extends FlatSpec with Matchers with Either
     }
   }
 
-  private def getMaxMinDates(dates: List[LocalDate]): (LocalDate, LocalDate) = {
+  private def getMaxDate(dates: List[LocalDate]): LocalDate = {
     val sortedDates = dates.sortWith((date1, date2) => date1.isBefore(date2))
-    (sortedDates.head, sortedDates.last)
+    sortedDates.last
   }
 
   private def getDatesBetween(startDate: LocalDate, endDate: LocalDate): List[LocalDate] = {
@@ -82,13 +42,10 @@ class SubscriptionDataIntegrationTest extends FlatSpec with Matchers with Either
   }
 
 
-  private def getDatesToTest(expectedIssues: List[IssueData]) = {
-    val (minDate, maxDate) = getMaxMinDates(expectedIssues.map(_.issueDate))
+  private def getDatesToTest(startDate: LocalDate, expectedIssues: List[IssueData]) = {
+    val maxDate = getMaxDate(expectedIssues.map(_.issueDate))
 
-    val datesCoveringExpectedIssuesPeriod = getDatesBetween(minDate, maxDate)
-
-    val dateBeforeExpectedIssues = minDate.minusDays(1)
-
-    (dateBeforeExpectedIssues :: datesCoveringExpectedIssuesPeriod)
+    getDatesBetween(startDate.minusDays(1), maxDate)
   }
+
 }
